Add check for payments section to only display implemented payment gateways (#1124)
Previously:
PaymentMethods was trying to import components even if those were not
defined in the `paymentMethodToComponent` mapping. It was leading to
errors if API returned gateway ID not known to the storefront.

After this change:
List of payment gateways is filtered by known implementations, and only
those will be displayed.